refactor(discord-bot): separate bot entrypoints
