+++ apps/discord-bot/src/discord/commands/publish.rs
@@ -1,7 +1,7 @@
+use crate::Error;
 use crate::alm::api::Client;
 use crate::alm::api::client::PublishMapError;
-use crate::types::GeoGuessrMap;
-use crate::{Context, Error};
+use crate::discord::{Context, geoguessr::GeoGuessrMap};
 use log::error;
 
 // make sure to restrict command usage via server permissions
