--- apps/discord-bot/src/types/mod.rs
@@ -1,3 +1,0 @@
-pub mod geoguessr;
-
-pub use geoguessr::GeoGuessrMap;
